fix(hooks): delegation-watch review fixes — skip scan once spent, sidechain guard, loud shape-drift skip, int thresholds, template parity
reviewer-correctness findings on ee4d296: (1) spent marker now short-circuits
the whole-file transcript scan (Stop fires per turn, not once per session);
(2) isSidechain entries excluded defensively (verified separate agent-*.jsonl
files today); (3) assistant entries with zero content arrays record
check_skipped format_unparseable instead of dying silently (P5); (4) marker
compare guards absent session_id; (5) thresholds gain .int(); (6)
delegation_watch block added to templates/default-config.json for parity.
Suites: 8/8 delegation, 100/100 hooks-full, 50/50 schemas, 800 pass total.

+++ hooks/h19-clear-session.mjs
@@ -4618,8 +4618,8 @@ var configSchema = external_exports.object({
   // >= min_hand_work AND (Task/Agent dispatches) <= max_dispatches. Defaults
   // calibrated on the measured 2026-08-10 incident (~23 hand-reads, 0 dispatches).
   delegation_watch: external_exports.object({
-    min_hand_work: external_exports.number().positive().default(15),
-    max_dispatches: external_exports.number().nonnegative().default(0)
+    min_hand_work: external_exports.number().int().positive().default(15),
+    max_dispatches: external_exports.number().int().nonnegative().default(0)
   }).default({}),
   // §7.2 model + effort defaults (tunable config, not architecture).
   // Hard rule encoded here as data: no xhigh/max for subagents except
